chore: refactor to modular cli
See: https://github.com/yargs/yargs/blob/master/docs/advanced.md#commanddirdirectory-opts
