--- examples/schema.json
@@ -1,11 +1,0 @@
-{
-  "$id": "http://example.com/schemaV1.json",
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "type": "object",
-  "properties": {
-    "key1": {
-      "type": "number"
-    }
-  },
-  "required": ["key1"]
-}
