--- examples/wrapped-certs/example.0.json
@@ -1,35 +1,0 @@
-{
-  "version": "open-attestation/3.0",
-  "data": {
-    "reference": "bb580cf5-87f2-42c8-b0cf-e9d445db336d:string:ABCXXXXX00",
-    "name": "84f3857b-40f7-41e7-8f71-cb8008fab163:string:Certificate of whatever",
-    "template": {
-      "name": "25d8bf44-0470-41cf-8ada-39d72c7364df:string:CUSTOM_TEMPLATE",
-      "type": "b82ae6ef-1f01-4714-b653-72debeea1839:string:EMBEDDED_RENDERER",
-      "url": "bfea6de8-db78-46f7-884b-ebbd1339e550:string:http://localhost:3000/rederer"
-    },
-    "validFrom": "b99a84c7-3871-4d33-996a-ff03a221371b:string:2018-08-30T00:00:00+08:00",
-    "proof": {
-      "type": "710ee6cf-5cea-41b9-a71b-106ad73460a9:string:OpenAttestationSignature2018",
-      "method": "07d987ad-97b8-4f9e-b06e-a3691a31ae03:string:TOKEN_REGISTRY",
-      "value": "509cc6a9-3c7d-4214-be79-d2b5fe4a5731:string:0xb53499ee758352fAdDfCed863d9ac35C809E2F20"
-    },
-    "issuer": {
-      "id": "84cf6e77-ed24-416d-bf1a-bd2a88058211:string:https://example.com",
-      "name": "90ded788-2f16-4b11-bfd4-21c58445730e:string:Issuer name",
-      "identityProof": {
-        "type": "283b3e84-b0e5-4dcb-a925-0036fa1f78ea:string:DNS-TXT",
-        "location": "75a891d6-9ba9-4d5e-9877-efe966dcd901:string:some.io"
-      }
-    }
-  },
-  "privacy": {
-    "obfuscatedData": []
-  },
-  "signature": {
-    "type": "SHA3MerkleProof",
-    "targetHash": "25d71e606ca2b56c1a5345101264626c9c7b89f124f21182b27e46ec208690ae",
-    "proof": [],
-    "merkleRoot": "25d71e606ca2b56c1a5345101264626c9c7b89f124f21182b27e46ec208690ae"
-  }
-}
