+++ package.json
@@ -69,7 +69,7 @@
     "node-fetch": "^2.6.0",
     "signale": "^1.4.0",
     "tmp": "^0.0.33",
-    "yargs": "^11.0.0"
+    "yargs": "^15.3.1"
   },
   "license": "GPL-3.0",
   "babel": {
