+++ src/implementations/filter/filter.ts
@@ -2,12 +2,6 @@ import fs from "fs";
 import signale from "signale";
 import { obfuscateDocument, verifySignature } from "@govtechsg/open-attestation";
 
-interface FilterCommand {
-  source: string;
-  destination: string;
-  fields: string[];
-}
-
 export const filter = (input: string, output: string, fields: string[]): void => {
   const documentJson = JSON.parse(fs.readFileSync(input, "utf8"));
   const obfuscatedDocument = obfuscateDocument(documentJson, fields);
@@ -19,8 +13,4 @@ export const filter = (input: string, output: string, fields: string[]): void =>
     fs.writeFileSync(output, JSON.stringify(obfuscatedDocument, null, 2));
     signale.success(`Obfuscated document saved to: ${output}`);
   }
-};
-
-export const isFilterCommand = (args: any): args is FilterCommand => {
-  return args._[0] === "filter";
 };
